requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
nltk==3.7
pandas==1.5.2
requests==2.28.1
scikit_learn==1.1.3
